requirements.txt
aiogram==3.17
asyncpg
pillow
python-dotenv